Refactor deployment module for Identity Verification Hub V2.
